move news scripts to news-sidebar.js
